feature: 产品, 新闻支持序号排序，序号相同按创建日期排序

+++ ruoyi-admin/src/main/java/com/ruoyi/greatzc/domain/Product.java
@@ -43,7 +43,7 @@ public class Product extends BaseEntity {
      * 排序
      */
     @Excel(name = "排序")
-    private Long orderNo;
+    private Integer orderNo;
 
     /**
      * 标题
